testing, refactoring and documentation
